Reduce overhead of multi-type-ness (last Summer's SoC): 	Use uint32_t instead of size_t counts (halves size of ChanCount on 64-bit). 	Shift DataType values down to eliminate subtraction every index of a ChanCount or *Set.
Allow using DataType directly as an array index (prettier/terser).
Fix some mixed spaces/tabs in file comment headers.


git-svn-id: svn://localhost/ardour2/trunk@2082 d708f5d6-7413-0410-9779-e7cbd77b26cf

+++ libs/ardour/chan_count.cc
@@ -33,7 +33,7 @@ infinity_factory()
 	ChanCount ret;
 
 	for (DataType::iterator t = DataType::begin(); t != DataType::end(); ++t) {
-		ret.set(*t, SIZE_MAX);
+		ret.set(*t, UINT32_MAX);
 	}
 
 	return ret;
